generateParticles method

+++ header/CircularBeam.h
@@ -9,7 +9,7 @@
 
 class CircularBeam : public Beam{
 public:
-    CircularBeam(const Particle &refParticle_, double lambda_, size_t nbrMacroParticles) : Beam(refParticle_, lambda_,
+    CircularBeam(const Particle &refParticle_, size_t lambda_, size_t nbrMacroParticles) : Beam(refParticle_, lambda_,
             nbrMacroParticles) { generateParticles(); }
 
 private:
